requirements.txt
Django==3.2.7
django-cassandra-engine==1.6.3
python-decouple==3.4
